Duration correction

+++ esprit1718b4eventmanagement/esprit1718b4eventmanagement-service/src/main/java/tn/esprit/b4/esprit1718b4eventmanagement/manufacturingservices/ManufacturingPlanningServiceLocal.java
@@ -1,6 +1,7 @@
 package tn.esprit.b4.esprit1718b4eventmanagement.manufacturingservices;
 
 import java.util.Date;
+import java.util.List;
 
 import javax.ejb.Local;
 
@@ -11,7 +12,7 @@ import tn.esprit.b4.esprit1718b4eventmanagement.utilities.IGenericDAO;
 
 @Local
 public interface ManufacturingPlanningServiceLocal extends IGenericDAO<ManufacturingPlanning> {
-	public long manufacturingDuration(Article article,int quantity);
+	public int manufacturingDuration(Article article,int quantity);
 	public Date endingManufacturingDate(Date startingDate,long duration);
