+++ esprit1718b4eventmanagement/esprit1718b4eventmanagement-service/src/main/java/tn/esprit/b4/esprit1718b4eventmanagement/manufacturingservices/ManufacturingPlanningServiceRemote.java
@@ -1,6 +1,7 @@
 package tn.esprit.b4.esprit1718b4eventmanagement.manufacturingservices;
 
 import java.util.Date;
+import java.util.List;
 
 import javax.ejb.Remote;
 
@@ -12,7 +13,7 @@ import tn.esprit.b4.esprit1718b4eventmanagement.utilities.IGenericDAO;
 @Remote
 public interface ManufacturingPlanningServiceRemote extends IGenericDAO<ManufacturingPlanning> {
 
-	public long manufacturingDuration(Article article,int quantity);
+	public int manufacturingDuration(Article article,int quantity);
 	public Date endingManufacturingDate(Date startingDate,long duration);
 	
 }
